export AuthenContinue and return it from ServerSession GetXXX functions.
ServerSession.sendReply() was returning an error when a session is
aborted in an authentication continue packet. Its not really an error,
so return the message and abort boolean as an AuthenContinue struct.
(we combine the user_msg and data packet fields as Message as from the
rfc it appears only one is relevant dependant on the abort flag)